Support nested objects within attributes when Marshaling
Unmarshaling an attribute that is a struct or struct pointer that is
decorated with jsonapi tags has historically worked as expected, but,
curiously, marshaling did not and required the use of `json` tags
instead, making struct reuse difficult for both input and output
object attributes.

Now, you can specify a struct or struct pointer as an attribute and
it will be marshaled into the correct keys.

Note that this implemenation does not yet support slices of structs or
struct pointers.